Number型でも四則演算可能にした
