+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod error;
 pub mod parser;
+pub mod number;
